Zoneminder added as testing only
Zoneminder was added since blog posts mentioned the prefix addition didn't work to properly route to ZM portal, however, it was found that it only fails the first time.

If you open zoneminder locally, accept the privacy notice or whatever is displayed the first time, then you will be able to properly enter from your subdomain.

This instance of someminder is the initial testing. Proper configuration is needed to move values into variables and point to correct locations.